Implemented local patientfiles
For phase 1 pilot tests, we need to be able to run without a server. To do that, I added a preference switch you can get at through "settings" on the iPad which switches on or off the use of a remote bonjour advertised server. Default is "NO", meaning local use without a server.
Added in the code in the PatientContextDB to load and save to local file instead of server if this switch is set to NO, else as usual to the server, still with HTTP.