Fix for crash on tapping in new field while old field was selected
It's complicated… Had to stop using dequeue and keep a pointer to a
table view cell in the IDRItem itself. But this stuff still crashes on
iOS 5.0 due to something that must be a bug in iOS 5 (beta). You get a
crash on UITextFieldEditor which turns out to be over released. This
does not happen on iOS 4.2. I also removed a bunch of NSLog()
statements. Added a mechanism for endEditing on table view, so had to
add a property for pointing to the parent table view in the ItemCell. I
also created custom UITableView and UITableCellView children, but I
didn't need to. Left them there anyway, so it's easier to override stuff
in the future.