Preferred language
Now the system chooses selection prompts and even value strings (which internally
are also a kind of prompt) according to the preferred language settings in the
system. It still doesn't update the preexisting values in fields according to
changed system settings. It uses the ISO 639 two letter language codes, so
issue templates need to use those, which are typically 'en', 'sv', 'nl' etc.
Note that language codes already in the data dictionary don't get fixed up yet,
so to demo this, you need to create a new patient.